refactored sidebar secondary section
